Added timeout parameter for API objects
By default there is no timeout on requests to the InstagramAPI. This can
cause unexpected behavior if a system is not anticipating a request to
block indefinitely.

This patch stops short of imposing a default timeout value, but does
expose the ability to specify a timeout value upon creating an api object.

The README has also been updated.